Added stub to read users from the web servce.
Since the first stop is reading all the users from the database, I've
added connection stub code to the method which calls the read users PHP
script.

A connection is established using GET and data is read from the server's
input stream. Later, the stub will be filled with code to actually parse
the server's response and create objects.