Read and parse data returned by user read and create scripts.
Added logic to send a request to read/create users, get the response and
parse the resulting user information if the query was successful; the
respective methods return appropriate values based on success/failure.

Code to generate a JSON object to be sent to the script (containing the
query parameters) remains to be added.