Modified creation of POST request.
The POST response received from the PHP script has the actual response
start after 2 lines, one of which is a single space and other is an empty
string. It needs to be debugged to find out why such lines are being
generated, but for now they are read, discarded and the third line is
used.