Adapter for the seller manager book list
Adapter implementation for displaying the list of books posted by the currently logged in user is now added. The layout for the display of each book is presently highly minimalistic, and will be improved in the later commits. The entire backend logic for getting the books and displaying them has been tested and works fine.